got carousel to not be too small or too big
